code/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
cytoolz==0.11.2
eth-abi==3.0.0
eth-hash==0.3.2
eth-typing==3.0.0
eth-utils==2.0.0
fonttools==4.31.1
frozenlist==1.3.0
idna==3.3
kiwisolver==1.4.0
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.3
packaging==21.3
parsimonious==0.8.1
Pillow==9.0.1
pycryptodome==3.14.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
toolz==0.11.2
urllib3==1.26.9
yarl==1.7.2
